Ensure we use ints

+++ zoidberg/rzline.py
@@ -184,7 +184,7 @@ class RZline:
         n = len(self.R)
         theta = np.remainder(theta, 2.0 * pi)
         dtheta = 2.0 * np.pi / n
-        ind = np.trunc(theta / dtheta)
+        ind = np.trunc(theta / dtheta).astype(int)
         rem = np.remainder(theta, dtheta)
         indp = (ind + 1) % n
         return (rem * self.R[indp] + (1.0 - rem) * self.R[ind]), (
